[FIX] make logout reply of app does its job (logout)
processing logout in NavView is not good. make logout reply does its job
