+++ pong/app/javascript/srcs/views/NavView.js
@@ -1,8 +1,6 @@
 import Radio from 'backbone.radio';
-import $ from 'jquery/src/jquery';
 import common from '../common';
 import template from '../templates/NavView.html';
-import auth from '../utils/auth';
 
 function link(event) {
   const href = event.target.getAttribute('href');
@@ -13,17 +11,7 @@ function link(event) {
 
 function logout(event) {
   event.preventDefault();
-  const data = {};
-  data[auth.getTokenKey()] = auth.getTokenValue();
-  $.ajax({
-    type: 'DELETE',
-    url: '/sign_out',
-    data,
-    success() {
-      Radio.channel('app').request('logout');
-      Radio.channel('route').trigger('route', 'login');
-    },
-  });
+  Radio.channel('app').request('logout');
 }
 
 const NavView = common.View.extend({
